+++ src/main/java/org/greports/engine/ReportLoaderResult.java
@@ -6,8 +6,10 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 public class ReportLoaderResult implements Serializable {
@@ -16,24 +18,52 @@ public class ReportLoaderResult implements Serializable {
 
     private final Map<Class<?>, List<?>> results = new HashMap<>();
     private final Map<Class<?>, List<ReportLoaderError>> errors = new HashMap<>();
+    private final Map<Class<?>, Set<Integer>> rowsWithErrors = new HashMap<>();
 
     protected <T> void addResult(Class<T> clazz, List<T> list) {
         results.put(clazz, list);
     }
 
     protected <T> void addError(Class<T> clazz, Cell cell, String columnTitle, String errorMessage, final Serializable errorValue) {
-        errorsCheckClass(clazz);
-        errors.get(clazz).add(new ReportLoaderError(cell, columnTitle, errorMessage, errorValue));
+        addError(clazz, new ReportLoaderError(cell, columnTitle, errorMessage, errorValue));
     }
 
     protected <T> void addError(Class<T> clazz, String sheetName, Integer rowIndex, Integer columnIndex, String columnTitle, String errorMessage, final Serializable errorValue) {
+        addError(clazz, new ReportLoaderError(sheetName, rowIndex, columnIndex, columnTitle, errorMessage, errorValue));
+    }
+
+    private <T> void addError(Class<T> clazz, ReportLoaderError error) {
         errorsCheckClass(clazz);
-        errors.get(clazz).add(new ReportLoaderError(sheetName, rowIndex, columnIndex, columnTitle, errorMessage, errorValue));
+        errors.get(clazz).add(error);
+        rowsWithErrorsCheckClass(clazz);
+        rowsWithErrors.get(clazz).add(error.getRowIndex());
+    }
+
+    private <T> void rowsWithErrorsCheckClass(Class<T> clazz) {
+        if (!rowsWithErrors.containsKey(clazz)) {
+            rowsWithErrors.put(clazz, new HashSet<>());
+        }
     }
 
     @SuppressWarnings("unchecked")
     public <T> List<T> getResult(Class<T> clazz) {
         return ((List<T>) results.getOrDefault(clazz, new ArrayList<>()));
+    }
+
+    public <T> List<T> getResultWithoutErrors(Class<T> clazz) {
+        List<T> toReturn = new ArrayList<>();
+        Set<Integer> rowsWithError = rowsWithErrors.getOrDefault(clazz, new HashSet<>());
+        List<T> result = getResult(clazz);
+        if(rowsWithError.size() == 0) {
+            return result;
+        }
+
+        for (int i = 0; i < result.size(); i++) {
+            if(!rowsWithError.contains(i + 1)) {
+                toReturn.add(result.get(i));
+            }
+        }
+        return toReturn;
     }
 
     public <T> List<ReportLoaderError> getErrors(Class<T> clazz) {