+++ src/main/java/org/greports/utils/Translator.java
@@ -1,5 +1,6 @@
 package org.greports.utils;
 
+import java.text.MessageFormat;
 import java.util.Map;
 
 public class Translator {
@@ -13,13 +14,9 @@ public class Translator {
         this(new TranslationsParser(locale, translationDir, translationFileExt));
     }
 
-    public String translate(String key, String... params){
+    public String translate(String key, Object... params){
         if(key != null) {
-            String text = translations.getOrDefault(key, key).toString();
-            for (int i = 0; i < params.length; i++) {
-                text = text.replaceAll("\\{" + i + "}", params[i]);
-            }
-            return text;
+            return MessageFormat.format(translations.getOrDefault(key, key).toString(), params);
         }
         return null;
     }