+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [3.2.5] - 2021-05-11
+
+### Added
+- ReportLoaderResult. Added getResultWithoutErrors method.
+- NotEmptyValidator added.
+### Changed
+- Translator. The logic of method Translator.translate was migrated to MessageFormat.format method
+
+## [3.2.4] - 2021-04-21
+
+### Fixed
+- Fixed getConfigurator method.
+
 ## [3.2.2] - 2021-03-23
 
 ### Added